+++ src/nakamori/EditMember.java
@@ -225,9 +225,22 @@ public class EditMember {
   		Select.selectMember( );
   		Scanner sc = new Scanner(System.in) ;
   		System.out.println("\n～以上が現在の従業員データです～\n");
-  		System.out.print( "削除する従業員番号を入力してください。：" );
-  		String code = sc.nextLine();
   		
+  		String code =null;
+		while(true) {
+		System.out.print( "削除する番号を入力してください。：" );
+		code = sc.nextLine();
+		
+		if(Select.object_id_judge(4,code)) {
+			break;
+		}
+		else {
+			System.out.println("\n【エラー：存在しないデータ番号の入力】");
+			System.out.println("データ内に存在する番号を選んでください\n");
+			System.out.println("入力に戻ります");
+		}
+		}//while
+		
   		String sql = "DELETE FROM member WHERE member_no = ?;" ;
   		
   		try (
