+++ src/nakamori/EditReceipt.java
@@ -216,8 +216,22 @@ public class EditReceipt {
 		Select.selectReceipt( 6, null );
 		System.out.println("\n～以上が現在の注文データです～\n");
 		Scanner sc = new Scanner(System.in) ;
+		
+		String code =null;
+		while(true) {
 		System.out.print( "削除する番号を入力してください。：" );
-		String code = sc.nextLine();
+		code = sc.nextLine();
+		
+		if(Select.object_id_judge(1,code)) {
+			break;
+		}
+		else {
+			System.out.println("\n【エラー：存在しないデータ番号の入力】");
+			System.out.println("データ内に存在する番号を選んでください\n");
+			System.out.println("入力に戻ります");
+		}
+		}//while
+		
 		String sql = "DELETE FROM receipt WHERE main_id = ?;" ;
 		try ( 
 			Connection con = DriverManager.getConnection( url , user_name , password ) ;
